fix : edit categories

+++ routes/web.php
@@ -20,37 +20,13 @@ use App\Models\Review;
 |
 */
 
-Route::get(
-    '/admin/categories',
-    CategoryController::class . "@index"
-)->name("category.index");
-
-Route::get(
-    '/admin/categories/ajouter',
-    CategoryController::class . "@create"
-)->name("category.create");
-
-Route::get(
-    '/admin/categories/{category}',
-    CategoryController::class . "@show"
-)->name("category.show");
-
-Route::delete(
-    '/admin/categories/{category}',
-    CategoryController::class . "@destroy"
-)->name("category.destroy");
-
-Route::post(
-    '/admin/categories',
-    CategoryController::class . "@store"
-)->name("category.store");
-
-Route::get(
-    '/admin/categories/modifier/{category}',
-    CategoryController::class . "@edit"
-)->name("category.edit");
-
-Route::put('/admin/modifier/{category}', CategoryController::class . "@update")->name("category.update");
+Route::get('/admin/categories', CategoryController::class . "@index")->name("category.index");
+Route::get('/admin/categories/ajouter', CategoryController::class . "@create")->name("category.create");
+Route::get('/admin/categories/{category}', CategoryController::class . "@show")->name("category.show");
+Route::delete('/admin/categories/{category}', CategoryController::class . "@destroy")->name("category.destroy");
+Route::post('/admin/categories', CategoryController::class . "@store")->name("category.store");
+Route::get('/admin/categories/modifier/{category}', CategoryController::class . "@edit")->name("category.edit");
+Route::put('/admin/categories/modifier/{category}', CategoryController::class . "@update")->name("category.update");
 
 Route::get('/admin/catalogue', "App\Http\Controllers\ProductController@index")->name('product.index');
 Route::get('/admin/produit/modifier/{product}', "App\Http\Controllers\ProductController@edit")->name('product.edit');
